# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OurDAO Backend API",
    "description": "Off-chain indexer + read API for the OurDAO lending DAO on Stellar/Soroban",
    "version": "0.2.0"
  },
...
  "paths": {
...
    "/api/members/{address}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "address",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    }
  ],
...
}